% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet.R
\name{alphabet}
\alias{alphabet}
\title{Some images}
\usage{
alphabet()
}
\value{
A character vector of 24 urls pointing to an image.
}
\description{
Twenty-four urls pointing to an image.
}
\examples{
freewall(alphabet(), widths = 200)
}
