% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freewall.R
\name{freewall}
\alias{freewall}
\title{<Add Title>}
\usage{
freewall(
  images,
  widths = 100,
  width = "100\%",
  draggable = FALSE,
  animate = TRUE,
  cellW = 100,
  cellH = 100,
  delay = 0,
  fixSize = NULL,
  gutterX = 10,
  gutterY = 10,
  keepOrder = TRUE,
  rightToLeft = FALSE,
  bottomToTop = FALSE,
  elementId = NULL
)
}
\description{
<Add Description>
}
