% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{nature}
\alias{nature}
\title{Some images}
\usage{
nature()
}
\value{
A character vector of nine urls pointing to an image.
}
\description{
Nine urls pointing to an image.
}
\examples{
freewall(
  nature(), widths = 200, cellW = 200, cellH = "200", draggable = TRUE
)
}
